[MDU-34]: Add BigDecimalConverter.kt and update Dbo's and Dto's. Add remaining domain entities
* [MDU-34]: Add BigDecimalConverter.kt and update Dbo's

* [MDU-34]: Update Mappers.kt

* [MDU-34]: Update Mappers.kt and add remaining entities

* Merge branch 'develop' into feature/updated-data-layer

* [MDU-34]: Rename and Move BigDecimalConverter.kt

* [MDU-34]: Add null safety on AppPreferences.kt

* [MDU-34]: Rename Converters.kt funcs

* [MDU-34]: Change general imports

* [MDU-34]: Update naming

* [MDU-34]: Add Bd migration

* [MDU-34]: Refactor

* [MDU-34]: Refactor to implement alerts in CartItem object

* [MDU-34]: Update DB version


Approved-by: Antonio Esteban Lorenzo
Approved-by: Joan Cremades Meló

+++ Project/app/src/main/java/com/example/modu/data/dataSource/remote/cart/dto/PriceAlertItemDto.kt
@@ -1,9 +1,9 @@
 package com.example.modu.data.dataSource.remote.cart.dto
 
-import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class PriceAlertItemDto(
     val productVariantId: Int,
-    val oldPrice: Double? = null,
-    val newPrice: Double? = null
+    val oldPrice: BigDecimal? = null,
+    val newPrice: BigDecimal? = null
 )
