+++ Project/app/src/main/java/com/example/modu/domain/entity/cart/PriceAlertItem.kt
@@ -1,7 +1,9 @@
 package com.example.modu.domain.entity.cart
 
+import java.math.BigDecimal
+
 data class PriceAlertItem(
     val productVariantId: Int,
-    val oldPrice: Long,
-    val newPrice: Long
+    val oldPrice: BigDecimal,
+    val newPrice: BigDecimal
 )
